Misc fixes
* Allow ATS with layout optimization
* Enable code paths in weak-cpu to be controlled by macros

+++ include/memfd.h
@@ -21,7 +21,7 @@ class MEMFD {
 private:
   size_t len;
   int ring_fd;
-  int pagesize;
+  long pagesize;
   size_t offset;
 #ifndef USE_MEMFD
   static std::string shm_prefix;
